CommitClass: first prototype
This is a very(!) rough draft of the CommitClass functionality described in
https://github.com/kubernetes/enhancements/pull/1078.

Some of the shortcuts taken in building this prototype:

1) No tests.
2) CommitClass struct defs live in
pkg/kubelet/commitclass/commitclass_manager.go, not in 'apis' somewhere.
This is because the 'node.k8s.io' API group isn't introduced until 1.14,
and in order to deploy internally I wanted to build on top of 1.12.x.
3) No safeguards against setting resources to zero.
4) No validation at all, really.
5) No exploration of how this functionality interacts with CPU
topology/CPU exclusivity.
6) CommitClass is a CRD living in `cluster/addons/commitclass`. Not at
all sure that's the right place for it.
7) No metrics indicating commit percentage.
8) No feature gates.

On the whole, this commit is entirely unsuitable for consumption by anyone
other than me, and I expect it to be nearly 100% rewritten before it
makes sense for wider use.