CommitClass: restrict commitsettings to whitelisted resources
This is an open design question, but one of the challenges of the
CommitClass feature is the potential to inflict major harm on the
cluster. That is amplified when the CommitClass API allows an operator
to over/under subscribe 'hard' resources like NICs, GPUs, hugepages, or
even pods.

As such, this commit restricts the set of resources CommitClass can
mutate to "the big three" of CPU, memory, and ephemeral storage. If
CommitClass gains traction as a feature, we can consider expanding its
scope further down the line.